rename css-injection-utils to css-rules, refactor css-rules
